Refactor: 도메인별 패키지 격리 및 UseCase/Repository 계층 구조 재배치
