+++ src/main/java/backend/mossy/boundedContext/cash/app/usecase/sync/CashSyncUserUseCase.java
@@ -1,8 +1,8 @@
-package backend.mossy.boundedContext.cash.app.usecase;
+package backend.mossy.boundedContext.cash.app.usecase.sync;
 
 import backend.mossy.boundedContext.cash.domain.user.CashUser;
-import backend.mossy.boundedContext.cash.out.CashUserRepository;
-import backend.mossy.boundedContext.cash.out.UserWalletRepository;
+import backend.mossy.boundedContext.cash.out.user.CashUserRepository;
+import backend.mossy.boundedContext.cash.out.user.UserWalletRepository;
 import backend.mossy.global.eventPublisher.EventPublisher;
 import backend.mossy.shared.cash.event.CashUserCreatedEvent;
 import backend.mossy.shared.member.dto.event.UserDto;
